Consolidate duplicated logic on reset tools
